wip utilize concerns
